introduce utils.twitter_login and use it in streaming and get_user_ids
