+++ streaming.py
@@ -11,6 +11,7 @@ from urllib3.exceptions import ProtocolError
 from tweepy.streaming import StreamListener
 
 import config as c
+import utils
 
 class StdOutListener(StreamListener):
     """
@@ -54,17 +55,7 @@ def run():
     }
 
     listener = StdOutListener(database)
-
-    auth = tweepy.OAuthHandler(config["consumer_key"], config["consumer_secret"])
-    auth.set_access_token(config["access_token"], config["access_token_secret"])
-
-    api = tweepy.API(auth)
-    # test authentication
-    try:
-        api.verify_credentials()
-        logging.info("authentification OK")
-    except:
-        logging.error("Error during authentication")
+    api = utils.twitter_login(conf)
 
     def signal_handler(*argv, **argh):
         database.close()
